requirements.txt
flask
beautifulsoup4
requests
openpyxl
pandas
lxml
python-dotenv
